Move CSV output to internal as preparations for v1.0.0
This was not moved before as a utility function `IsStringInSlice` used
by a now not maintained extension:
- https://github.com/Maksimall89/xk6-output-clickhouse

As that dependency doesn't make any sense it was planned to be moved
either way. But as the extension is now not maintained and has not merged
previous PRs to fix breaking changes, it is even more of a reason to
move forward.

Related to #4133
